implemented support for dummy atoms
Dummy atoms such as the MW dummy in TIP4P etc. are now treated as bonded to the first atom of the preceding molecule tree even though they are not involved in any bond lists. Thus they are correctly included in the unwrapping of periodic boundary conditions

+++ MDA_unwrap_PBC/unwrap.py
@@ -27,7 +27,7 @@ clib.buildTrees.argtypes = [
     ct.c_int32,
     np.ctypeslib.ndpointer(dtype=np.int32, ndim=1, flags='C_CONTIGUOUS'),
     np.ctypeslib.ndpointer(dtype=np.int32, ndim=2, flags='C_CONTIGUOUS'),
-    np.ctypeslib.ndpointer(dtype=np.int32, ndim=1, flags='C_CONTIGUOUS')
+    np.ctypeslib.ndpointer(dtype=np.float32, ndim=1, flags='C_CONTIGUOUS')
 ]
 #define return type of function 'buildTrees' in imported library 'clib'
 clib.buildTrees.restype = ct.c_int32
